config: likedStoreCard 컴포넌트 기능 storeCard 컴포넌트에 합치기
- storeCardInfo 타입 수정
- likedCard 타입 추가 기본 false
